+++ threedigrid_builder/interface/gridadmin.py
@@ -158,10 +158,9 @@ class GridAdminOut(OutputInterface):
             self.write_quadtree(grid.quadtree_stats)
         self.write_nodes(grid.nodes)
         self.write_nodes_embedded(grid.nodes_embedded)
-        self.write_lines(grid.lines)
+        self.write_lines(grid.lines, grid.cross_sections)
         self.write_pumps(grid.pumps)
-        if grid.cross_sections.tables is not None:
-            self.write_cross_sections(grid.cross_sections)
+        self.write_cross_sections(grid.cross_sections)
         self.write_levees(grid.levees)
         self.write_breaches(grid.breaches)
 
@@ -343,15 +342,15 @@ class GridAdminOut(OutputInterface):
         # unknown
         self.write_dataset(group, "sumax", np.full(shape, np.nan, dtype=np.float64))
 
-    def write_lines(self, lines):
+    def write_lines(self, lines, cross_sections):
         """Write the "lines" group in the gridadmin file
 
         Raises a ValueError if it exists already.
 
         Notes:
             Some datasets were 64-bit integers, but now they are saved as 32-bit integers.
-            The following datasets were added: ds1d, dpumax, flod, flou, cross1, cross2,
-            cross_weight
+            The following datasets were added: ds1d, dpumax, flod, flou, cross1,
+            cross2, cross_weight
             For floats (double) we use NaN instead of -9999.0 to denote empty.
 
         Args:
@@ -405,8 +404,17 @@ class GridAdminOut(OutputInterface):
         )
         self.write_dataset(group, "flod", lines.flod)
         self.write_dataset(group, "flou", lines.flou)
-        self.write_dataset(group, "cross1", increase(lines.cross1))
-        self.write_dataset(group, "cross2", increase(lines.cross2))
+
+        cross1 = lines.cross_id1.copy()
+        cross1[cross1 != -9999] = np.digitize(
+            cross1[cross1 != -9999], cross_sections.content_pk, right=True
+        )
+        cross2 = lines.cross_id2.copy()
+        cross2[cross2 != -9999] = np.digitize(
+            cross2[cross2 != -9999], cross_sections.content_pk, right=True
+        )
+        self.write_dataset(group, "cross1", increase(cross1))
+        self.write_dataset(group, "cross2", increase(cross2))
         self.write_dataset(group, "frict_type1", lines.frict_type1)
         self.write_dataset(group, "frict_type2", lines.frict_type2)
         self.write_dataset(group, "frict_value1", lines.frict_value1)
@@ -483,7 +491,7 @@ class GridAdminOut(OutputInterface):
         )
 
     def write_cross_sections(self, cross_sections: CrossSections):
-        if cross_sections is None or cross_sections.tables is None:
+        if len(cross_sections) == 0:
             return
         group = self._file.create_group("cross_sections")